+++ kaithem/src/vue/ScriptEditor.vue
@@ -1,19 +1,3 @@
----
-allow-origins: ['*']
-allow-xss: false
-auto-reload: false
-auto-reload-interval: 5.0
-dont-show-in-index: true
-mimetype: text/html
-no-header: true
-no-navheader: false
-require-method: [GET, POST]
-require-permissions: []
-resource-timestamp: 1596593926736481
-resource-type: page
-template-engine: none
-
----
 <style scoped>
 
         .event {